Move test to a better place
